Add auth with nicknames to chat

+++ portfolio/src/main/webapp/script.js
@@ -26,63 +26,64 @@ function loadMessages() {
 
 /** Validate the form and add warning if something wrong. */
 function processMessageForm() {
-  document.getElementById('warn-container').innerText = "";
+  document.getElementById("message-warn-container").innerText = "";
   
-  var name = document.getElementById('name').value;
-  var text = document.getElementById('text').value;
+  var text = document.getElementById("text").value;
 
-  var warning = processWarning(name, text);
-  if (warning.length != 0) {
-    document.getElementById('warn-container').innerText = warning;
+  if (text == null || text.trim().length == 0) {
+    document.getElementById("message-warn-container").innerText = "Please write a message.\n";
     return false;
   }
-
   return true;
 }
 
-/** Returns the warning if the message is empty or name is anonym. Return "" if everything OK. */
-function processWarning(name, text) {
-  var warning = "";
-  if (name.trim().length == 0) {
-    warning += "Please choose a name.\n";
+function processNicknameForm() {
+  document.getElementById("nickname-warn-container").innerText = "";
+  
+  var nickname = document.getElementById("nickname").value;
+
+  if (nickname == null || nickname.trim().length == 0) {
+    document.getElementById("nickname-warn-container").innerText = "Please write a nickname.\n";
+    return false;
   }
-  if (text.trim().length == 0) {
-    warning += "Please write a message.\n";
-  }
-  return warning;
+  return true;
 }
 
 /** Creates an element that represents a message, including its delete button. */
 function createMessageElement(message) {
-  const messageElement = document.createElement('li');
+  const messageElement = document.createElement("li");
 
-  const timeElement = document.createElement('span');
-  timeElement.setAttribute('class', 'timestamp');
+  const timeElement = document.createElement("span");
+  timeElement.setAttribute("class", "timestamp");
   timeElement.innerText = formatTimestamp(message.timestamp);
 
-  const nameElement = document.createElement('span');
-  nameElement.setAttribute('class', 'name');
-  nameElement.innerText = message.name;
+  const nameElement = document.createElement("span");
+  nameElement.setAttribute("class", "nickname");
+  nameElement.innerText = message.nickname;
 
-  const textElement = document.createElement('span');
-  textElement.setAttribute('class', 'text');
+  const textElement = document.createElement("span");
+  textElement.setAttribute("class", "text");
   textElement.innerText = message.text;
+  
+  fetch("/get-uid").then(response => response.json()).then((current_uid) => {
+    console.log(current_uid, message.uid)
+    if (message.uid == current_uid) {
+      const deleteButtonElement = document.createElement("button");
+      deleteButtonElement.setAttribute("class", "delete");
+      deleteButtonElement.innerHTML="<img src=\"images/trash.svg\">";
+      deleteButtonElement.addEventListener("click", () => {
+        deleteMesage(message);
 
-  const deleteButtonElement = document.createElement('button');
-  deleteButtonElement.setAttribute('class', 'delete');
-  deleteButtonElement.innerHTML="<img src=\"images/trash.svg\">";
-  deleteButtonElement.addEventListener('click', () => {
-    deleteMesage(message);
-
-    // Remove the message from the DOM.
-    messageElement.remove();
+        // Remove the message from the DOM.
+        messageElement.remove();
+      });
+      messageElement.appendChild(deleteButtonElement);
+    }
+    messageElement.appendChild(timeElement);
+    messageElement.appendChild(nameElement);
+    messageElement.appendChild(textElement);
   });
 
-  messageElement.appendChild(deleteButtonElement);
-  messageElement.appendChild(timeElement);
-  messageElement.appendChild(nameElement);
-  messageElement.appendChild(textElement);
-  
   return messageElement;
 }
 
@@ -94,14 +95,60 @@ function formatTimestamp(timestamp) {
 /** Tells the server to delete the message. */
 function deleteMesage(message) {
   const params = new URLSearchParams();
-  params.append('id', message.id);
-  fetch('/delete-message', {method: 'POST', body: params});
+  params.append("id", message.id);
+  fetch("/delete-message", {method: "POST", body: params});
   loadMessages();
 }
 
 function loadCV() {
-  const CVContainer = document.getElementById('cv-container');
+  const CVContainer = document.getElementById("cv-container");
   CVContainer.innerText = "Opening...";
+}
+
+function loadNicknameForm() {
+  fetch("/is-logged-in").then(response => response.json()).then((isLoggedIn) => {
+    if (!isLoggedIn) {
+      document.getElementById("nickname-container").hidden = true;
+      fetch("/get-login-url").then(response => response.json()).then((URL) => {
+        document.getElementById("log-container").innerHTML = "<p>Want to choose a nickname? Please, <a href=" + URL + ">login</a></p>";
+      });
+    } else {
+      document.getElementById("nickname-container").hidden = false;
+      fetch("/get-logout-url").then(response => response.json()).then((URL) => {
+        document.getElementById("log-container").innerHTML = "<p>You can <a href=" + URL + ">logout</a></p>";
+      });
+      
+      fetch("/get-nickname").then(response => response.json()).then((nickname) => {
+        if (nickname != "" && nickname != null) {
+          document.getElementById("nickname").value = nickname;
+        }
+      });
+    }
+  });
+}
+
+function loadMessageForm() {
+  fetch("/is-logged-in").then(response => response.json()).then((isLoggedIn) => {
+    if (!isLoggedIn) {
+      document.getElementById("message-container").hidden = true;
+      fetch("/get-login-url").then(response => response.json()).then((URL) => {
+        document.getElementById("log-container").innerHTML = "<p>Want to send a message? Please, <a href=" + URL + ">login</a></p>";
+      });
+    } else {
+      document.getElementById("message-container").hidden = false;
+      fetch("/get-logout-url").then(response => response.json()).then((URL) => {
+        document.getElementById("log-container").innerHTML = "<p>You can <a href=" + URL + ">logout</a></p>";
+      });
+      
+      fetch("/get-nickname").then(response => response.json()).then((nickname) => {
+        if (nickname != "" && nickname != null) {
+          document.getElementById("nickname-container").innerHTML = "<p><span class=\"nickname\">" + nickname + "</span> Change <a href=\"/nickname.html\">here</a>.</p>";
+        } else {
+          document.getElementById("nickname-container").innerHTML = "<p>You don't have a nickname. Choose <a href=\"/nickname.html\">here</a>.</p>";
+        }
+      });
+    }
+  });
 }
 
 /**
@@ -109,8 +156,8 @@ function loadCV() {
  */
 
 function getData() {
-  fetch('/data').then(response => response.json()).then((comment) => {
-    document.getElementById('data-container').innerText = comment;
+  fetch("/data").then(response => response.json()).then((count) => {
+    document.getElementById("data-container").innerText = count;
   });
 }
 
@@ -118,7 +165,7 @@ function getData() {
  * Adds a random placeholder.
  */
 function getRandomPlaceholder(elementName) {
-  const placeholders = ['I love cheesecake jelly-o sweet.', 'I love gingerbread bear claw marshmallow toffee cookie donut marzipan jelly-o.', 'Cupcake jujubes donut croissant I love.'];
+  const placeholders = ["I love cheesecake jelly-o sweet.", "I love gingerbread bear claw marshmallow toffee cookie donut marzipan jelly-o.", "Cupcake jujubes donut croissant I love."];
 
   // Pick a random placeholder.
   const placeholder = placeholders[Math.floor(Math.random() * placeholders.length)];
@@ -131,11 +178,11 @@ function getRandomPlaceholder(elementName) {
  * Adds a random greeting to the page.
  */
 function addRandomGreeting() {
-  const greetings = ['Hello world!', '¡Hola Mundo!', '你好，世界！', 'Bonjour le monde!', 'Привет мир!', 'Բարեւ աշխարհ!'];
+  const greetings = ["Hello world!", "¡Hola Mundo!", "你好，世界！", "Bonjour le monde!", "Привет мир!", "Բարեւ աշխարհ!"];
   // Pick a random greeting.
   const greeting = greetings[Math.floor(Math.random() * greetings.length)];
 
   // Add it to the page.
-  const greetingContainer = document.getElementById('greeting-container');
+  const greetingContainer = document.getElementById("greeting-container");
   greetingContainer.innerText = greeting;
 }
